server/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
google-auth>=2.38
google-genai>=1.14
httpx>=0.28
